Config : make oracles and treasury configurable for future upgrades

+++ src/interfaces/IPyth.sol
@@ -16,21 +16,20 @@ struct PriceData {
 }
 
 interface IPyth {
-    
-    function getValidTimePeriod() external view returns (uint validTimePeriod);
-    
+    function getValidTimePeriod() external view returns (uint256 validTimePeriod);
+
     function getPrice(bytes32 id) external view returns (Price memory price);
     function getEmaPrice(bytes32 id) external view returns (Price memory price);
     function getPriceUnsafe(bytes32 id) external view returns (Price memory price);
-    
-    function getPriceNoOlderThan(bytes32 id, uint age) external view returns (Price memory price);
+
+    function getPriceNoOlderThan(bytes32 id, uint256 age) external view returns (Price memory price);
     function updatePriceFeeds(bytes[] calldata updateData) external payable;
-    
+
     function updatePriceFeedsIfNecessary(
         bytes[] calldata updateData,
         bytes32[] calldata priceIds,
         uint64[] calldata publishTimes
     ) external payable;
-    
-    function getUpdateFee(bytes[] calldata updateData) external view returns (uint feeAmount);
+
+    function getUpdateFee(bytes[] calldata updateData) external view returns (uint256 feeAmount);
 }
